Refactored a couple things to improve readability. Fixed a bug where the <pre> tags were not wrapping properly.
