+++ js/src/app/submission/SubmissionDetail.module.css
@@ -1,0 +1,12 @@
+.description {
+  min-width: 0;
+}
+
+.description > pre {
+  background-color: rgba(51, 48, 48, 0.8);
+  border-radius: 1rem;
+  padding: 2rem;
+  min-width: 0;
+  white-space: pre-wrap;
+  word-wrap: break-word;
+}
